+++ src/main/java/roomescape/payment/processor/toss/TossPaymentConfirmError.java
@@ -1,0 +1,7 @@
+package roomescape.payment.processor.toss;
+
+public record TossPaymentConfirmError(
+    String code,
+    String message
+) {
+}
